requirements.txt
fastapi
uvicorn[standard]
pydantic==1.*
python-dotenv
sqlmodel
sqlalchemy
uuid
asyncpg
sqlalchemy-utils
ujson
greenlet
logger
boto3
python-multipart
celery[redis]
Pillow
alembic
requests
requests-toolbelt
pyjwt
bcrypt
python-multipart
itsdangerous
stripe
passlib
librosa
psycopg2
aiofiles